\name{ShortPath-package}
\alias{ShortPath-package}
\alias{ShortPath}
\docType{package}
\title{
  Finding the shortest path of a graph
}
\description{
  ShortPath is an R package that allows determining the shortest path of a graph from a source vertex using two algorithms: Bellman-Ford algorithm and Dijkstra algorithm. In particular, this package makes it possible to determine the shortest path on a graph included in a square [0,0]*[1,1], which has as its starting vertex with coordinates (0,0) and as its arrival vertex with coordinates (1,1). An example of application of these algorithms is seam carving. It is an image resizing algorithm developed by Shai AVIDAN and Ariel SHAMIR, which resizes the image by removing the least energy pixel paths.
}
\author{
Chaïma Boughanmi, Faheemah Padavia, Pascale Zouein
}
\keyword{ package, Bellman Ford, Dijkstra, Seam Carving }
\seealso{
  https://fr.wikipedia.org/wiki/Algorithme_de_Bellman-Ford,
  https://fr.wikipedia.org/wiki/Algorithme_de_Dijkstra,
  https://fr.wikipedia.org/wiki/Seam_carving
}
